Backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
